Supress rye install log output when in debug mode
rye uses [whatthestell](https://github.com/zekroTJA/whattheshell)
[here](https://github.com/mitsuhiko/rye/blob/d950ca25b032d59846b78aba3dfb1284e3d04505/rye/src/cli/rye.rs#L392)
to determine what shell is being used. This internally uses `ps`
[here](https://github.com/zekroTJA/whattheshell/blob/b6290f96b5638314ac4547def37fe35985831d5d/src/infer/unix.rs#L52).
My guess is that this prints`error: process ID out of range`
to stderr when searching for the process which opened the shell
in the current process "tree"
because no shell is used when the `rye self install`
command is executed by this typescript action.