Move RYE_HOME from workspace home to runner home
Having it "suddenly appear" in the repo root confused some tools like `rye fmt`. We assume that the directory that gets created for the current workflow run is a stable path between workflow runs. If several repositories get checked out or some logic interacts on the workspace level this might lead to unintented behaviour. If this turns out to be a real issue we have to look into using TMPDIR and try to create stable paths there.