Move RYE_HOME from workspace home to runner home (#194)
Having it "suddenly appear" in the repo root confused some tools like `rye fmt`. We assume that the directory that gets created for the current workflow run is a stable path between workflow runs. If several repositories get checked out or some logic interacts on the workspace level this might lead to unintented behaviour. If this turns out to be a real issue we have to look into using TMPDIR and try to create stable paths there.

Several runs of the same repository on the same runner might run with different options. These must not interfere and thus the `.rye` folder must not be left over after the run. The only way to ensure that is to always run the `post` step of the action. Since there is (currently) no way to determine if the workflow succeeded or failed when the `post` step is started (it is possible to run or not run it) we will always save the cache. Since we are only caching `.rye` and `.venv` there should be no errors because of this.

Breaking: The cache will always be saved, even if the workflow failed.