Only save cache when run succeeded
Commit 9dddc348b6f40922e842cc66267f814bae8e105f disabled this with famous last words "there should be no errors because of this". This is wrong as a run can have some misconfigured or failing steps that create a partial or wrong .venv. This is then uploaded to the cache and each subsequent run finds that corrupted .venv in the cache and does not attempt create the full/correct version.
The initial reason for always running the post run step was to clean up the .rye home folder. I later learned that the only thing we need to clean up is the config.toml file. Upon first install of the .rye folder (for self-hosted runners. ephemeral runners always install) a backup of the clean config.toml file is created with the name config.toml.bak. If this file exists on subsequent it will be used to overwrite a possible changed config.toml. This only works for version 0.19.0 and later.